# code/backend/src/config/swagger.yaml
openapi: 3.0.0
info:
  title: API Reside
  version: 1.0.0
  description: API do Sistema de Gerenciamento de Condomínios Reside
servers:
  - url: https://reside-production.up.railway.app
    description: Servidor de Desenvolvimento Local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
          description: Uma mensagem descritiva do erro.
          example: 'Token inválido ou expirado.'
...
    Parking:
      type: object
      properties:
        id:
          type: integer
          example: 1
        resident_id:
          type: integer
          example: 5
        apartment:
          type: string
          example: "101A"
        condominium_id:
          type: integer
          example: 1
        location:
          type: string
          example: "Subsolo 1, Vaga 05"
        type:
          type: string
          enum: [diario, mensal]
          example: "diario"
        price:
          type: number
          format: float
          example: 25.00
        description:
          type: string
          example: "Vaga de fácil acesso, próxima ao elevador"
        status:
          type: string
          enum: [disponivel, reservado, indisponivel]
          example: "disponivel"
        is_covered:
          type: boolean
          example: true
        available_date:
          type: string
          format: date
          example: "2025-06-30"
        start_time:
          type: string
          example: "08:00"
        end_time:
          type: string
          example: "18:00"
        sunday:
          type: boolean
          example: false
        monday:
          type: boolean
          example: true
        tuesday:
          type: boolean
          example: true
        wednesday:
          type: boolean
          example: true
        thursday:
          type: boolean
          example: false
        friday:
          type: boolean
          example: false
        saturday:
          type: boolean
          example: false
        reserver_id:
          type: integer
          nullable: true
          example: 7
        reservation_expires_at:
          type: string
          format: date-time
          nullable: true
          example: "2025-06-30T18:00:00.000Z"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ParkingCreateRequest:
      type: object
      properties:
        resident_id:
          type: integer
          example: 5
        apartment:
          type: string
          example: "101A"
        condominium_id:
          type: integer
          example: 1
        location:
          type: string
          example: "Subsolo 1, Vaga 05"
        type:
          type: string
          enum: [diario, mensal]
          example: "diario"
        price:
          type: number
          format: float
          example: 25.00
        description:
          type: string
          example: "Vaga de fácil acesso, próxima ao elevador"
        is_covered:
          type: boolean
          example: true
        available_date:
          type: string
          format: date
          example: "2025-06-30"
        start_time:
          type: string
          example: "08:00"
        end_time:
          type: string
          example: "18:00"
        domingo:
          type: boolean
          example: false
        segunda:
          type: boolean
          example: true
        terca:
          type: boolean
          example: true
        quarta:
          type: boolean
          example: true
        quinta:
          type: boolean
          example: false
        sexta:
          type: boolean
          example: false
        sabado:
          type: boolean
          example: false
      required:
        - resident_id
        - apartment
        - condominium_id
        - location
        - type
        - price
        - is_covered
    ParkingRequestBody:
      type: object
      properties:
        resident_id:
          type: integer
          example: 7
      required:
        - resident_id
    ParkingRequestResponse:
      type: object
      properties:
        message:
          type: string
          example: "Solicitação de vaga realizada com sucesso"
        expiresAt:
          type: string
          format: date-time
          example: "2025-06-30T18:00:00.000Z"
    ExpiredReservationsResponse:
      type: object
      properties:
        message:
          type: string
          example: "2 reservas expiradas foram canceladas"
        canceledCount:
          type: integer
          example: 2
...
paths:
...
  /parkings:
    post:
      tags:
        - Estacionamento
      summary: Criar nova vaga de estacionamento
      description: Permite que um morador disponibilize uma vaga para aluguel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParkingCreateRequest'
      responses:
        '201':
          description: Vaga criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parking'
        '400':
          description: Dados inválidos ou limite de vagas por apartamento excedido
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
                
    get:
      tags:
        - Estacionamento
      summary: Listar todas as vagas
      description: Retorna todas as vagas de estacionamento cadastradas
      responses:
        '200':
          description: Lista de vagas retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Parking'
        '500':
          description: Erro interno do servidor
  /parkings/available:
    get:
      tags:
        - Estacionamento
      summary: Buscar vagas disponíveis
      description: Busca vagas disponíveis com base em filtros como tipo, data, horário, etc.
      parameters:
        - name: condominium_id
          in: query
          description: ID do condomínio
          schema:
            type: integer
        - name: location
          in: query
          description: Localização da vaga (para busca parcial)
          schema:
            type: string
        - name: type
          in: query
          description: Tipo de aluguel (diário ou mensal)
          schema:
            type: string
            enum: [diario, mensal]
        - name: max_price
          in: query
          description: Preço máximo
          schema:
            type: number
        - name: is_covered
          in: query
          description: Se a vaga é coberta
          schema:
            type: boolean
        - name: date
          in: query
          description: Data disponível (para vagas diárias)
          schema:
            type: string
            format: date
        - name: start_time
          in: query
          description: Hora de início (para vagas diárias)
          schema:
            type: string
        - name: end_time
          in: query
          description: Hora de fim (para vagas diárias)
          schema:
            type: string
        - name: weekdays
          in: query
          description: Dias da semana (para vagas mensais), formato JSON array de booleans
          schema:
            type: string
      responses:
        '200':
          description: Lista de vagas disponíveis
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Parking'
        '500':
          description: Erro interno do servidor
  /parkings/check-expired:
    post:
      tags:
        - Estacionamento
      summary: Verificar reservas expiradas
      description: Verifica e cancela automaticamente reservas que expiraram
      responses:
        '200':
          description: Reservas expiradas verificadas com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpiredReservationsResponse'
        '500':
          description: Erro interno do servidor
  /parkings/{id}:
    get:
      tags:
        - Estacionamento
      summary: Obter detalhes de uma vaga específica
      parameters:
        - name: id
          in: path
          required: true
          description: ID da vaga
          schema:
            type: integer
      responses:
        '200':
          description: Detalhes da vaga
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parking'
        '400':
          description: ID inválido
        '404':
          description: Vaga não encontrada
        '500':
          description: Erro interno do servidor
          
    put:
      tags:
        - Estacionamento
      summary: Atualizar uma vaga
      parameters:
        - name: id
          in: path
          required: true
          description: ID da vaga
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParkingCreateRequest'
      responses:
        '200':
          description: Vaga atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Parking'
        '400':
          description: Dados inválidos
        '404':
          description: Vaga não encontrada
          
    delete:
      tags:
        - Estacionamento
      summary: Remover uma vaga
      parameters:
        - name: id
          in: path
          required: true
          description: ID da vaga
          schema:
            type: integer
      responses:
        '204':
          description: Vaga removida com sucesso
        '400':
          description: ID inválido
        '404':
          description: Vaga não encontrada
        '500':
          description: Erro interno do servidor
  /parkings/{id}/request:
    post:
      tags:
        - Estacionamento
      summary: Solicitar uma vaga
      description: Permite que um morador solicite a reserva de uma vaga disponível
      parameters:
        - name: id
          in: path
          required: true
          description: ID da vaga
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ParkingRequestBody'
      responses:
        '200':
          description: Solicitação realizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ParkingRequestResponse'
        '400':
          description: ID inválido ou vaga já reservada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /parkings/{id}/cancel-request:
    post:
      tags:
        - Estacionamento
      summary: Cancelar uma solicitação de vaga
      description: Cancela uma reserva de vaga e a torna disponível novamente
      parameters:
        - name: id
          in: path
          required: true
          description: ID da vaga
          schema:
            type: integer
      responses:
        '200':
          description: Solicitação cancelada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Solicitação de vaga cancelada com sucesso"
        '400':
          description: ID inválido ou vaga não está reservada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /condos/{condominiumId}/parkings:
    get:
      tags:
        - Estacionamento
      summary: Listar vagas de um condomínio
      parameters:
        - name: condominiumId
          in: path
          required: true
          description: ID do condomínio
          schema:
            type: integer
      responses:
        '200':
          description: Lista de vagas do condomínio
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Parking'
        '500':
          description: Erro interno do servidor
  /condos/{condominiumId}/apartments/{apartment}/parkings:
    get:
      tags:
        - Estacionamento
      summary: Listar vagas de um apartamento
      parameters:
        - name: condominiumId
          in: path
          required: true
          description: ID do condomínio
          schema:
            type: integer
        - name: apartment
          in: path
          required: true
          description: Número do apartamento
          schema:
            type: string
      responses:
        '200':
          description: Lista de vagas do apartamento
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Parking'
        '500':
          description: Erro interno do servidor
  /residents/{residentId}/parkings:
    get:
      tags:
        - Estacionamento
      summary: Listar vagas de um residente
      parameters:
        - name: residentId
          in: path
          required: true
          description: ID do residente
          schema:
            type: integer
      responses:
        '200':
          description: Lista de vagas do residente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Parking'
        '500':
          description: Erro interno do servidor
  /residents/{residentId}/reserved-parkings:
    get:
      tags:
        - Estacionamento
      summary: Listar vagas reservadas por um residente
      parameters:
        - name: residentId
          in: path
          required: true
          description: ID do residente
          schema:
            type: integer
      responses:
        '200':
          description: Lista de vagas reservadas pelo residente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Parking'
        '500':
          description: Erro interno do servidor
...
  /parkings/{id}/rent:
    post:
      tags:
        - Estacionamento
      summary: Alugar uma vaga diretamente com notificação
      description: Permite que um morador alugue uma vaga disponível diretamente. Usa a mesma lógica da rota /request mas envia notificação automática para o proprietário da vaga.
      parameters:
        - name: id
          in: path
          required: true
          description: ID da vaga
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                renter_id:
                  type: integer
                  description: ID do morador que está alugando a vaga
              required:
                - renter_id
              example:
                renter_id: 123
      responses:
        '200':
          description: Vaga alugada com sucesso e proprietário notificado
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Vaga alugada com sucesso. O proprietário foi notificado."
                  expiresAt:
                    type: string
                    format: date-time
                    description: Data/hora de expiração da reserva
                    example: "2025-06-30T18:00:00.000Z"
        '400':
          description: Dados inválidos, vaga não disponível ou já reservada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                vagaNaoDisponivel:
                  summary: Vaga não está disponível
                  value:
                    message: "Vaga não está disponível"
                renterIdObrigatorio:
                  summary: Renter ID não fornecido
                  value:
                    message: "ID do locatário é obrigatório"
                idInvalido:
                  summary: ID da vaga inválido
                  value:
                    message: "ID da vaga deve ser um número válido"
                vagaExpirada:
                  summary: Vaga diária expirada
                  value:
                    message: "A data disponível da vaga já passou"
                residenteNaoEncontrado:
                  summary: Residente não encontrado
                  value:
                    message: "Residente não encontrado"
        '404':
          description: Vaga não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                vagaNaoEncontrada:
                  summary: Vaga não encontrada
                  value:
                    message: "Vaga não encontrada"
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                erroInterno:
                  summary: Erro interno
                  value:
                    message: "Erro interno do servidor"
...